+++ common/utils.go
@@ -1,6 +1,7 @@
 package common
 
 import (
+	"regexp"
 	"slices"
 	"strconv"
 	"unicode"
@@ -135,4 +136,10 @@ func SafeIntToUint64(val int) uint64 {
 	}
 
 	return uint64(val)
+}
+
+// IsUUID Validate if the string pass through is an uuid
+func IsUUID(s string) bool {
+	r := regexp.MustCompile("^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[1-5][a-fA-F0-9]{3}-[89abAB][a-fA-F0-9]{3}-[a-fA-F0-9]{12}$")
+	return r.MatchString(s)
 }