+++ components/transaction/internal/adapters/database/postgres/operation.postgresql.go
@@ -48,7 +48,7 @@ func (r *OperationPostgreSQLRepository) Create(ctx context.Context, operation *o
 	record := &o.OperationPostgreSQLModel{}
 	record.FromEntity(operation)
 
-	result, err := db.ExecContext(ctx, `INSERT INTO operation VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, $21, $22, $23, $24, $25) RETURNING *`,
+	result, err := db.ExecContext(ctx, `INSERT INTO operation VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19, $20, $21, $22, $23, $24) RETURNING *`,
 		record.ID,
 		record.TransactionID,
 		record.Description,